+++ backend/plugin/src/test/java/app/epistola/valtimo/web/rest/EpistolaGenerationResourceAuthorizationTest.java
@@ -71,7 +71,8 @@ class EpistolaGenerationResourceAuthorizationTest {
 
         resource = new EpistolaGenerationResource(pluginService, epistolaService,
                 previewService, retryFormService, jsonataMappingService,
-                documentService, objectMapper, authorizationService, operatonTaskService);
+                documentService, objectMapper, authorizationService, operatonTaskService,
+                mock(org.operaton.bpm.engine.RuntimeService.class));
     }
 
     private PreviewRequest validPreviewRequest() {